Mostly adds a bitreader to the equation
The bitreader is pretty poorly tested. It looks like
it's working fine so far.

It's addition is necessary in order to properly parse
the sub-frame. Each frame is aligned to the byte boundary,
but each sub-frame is not, so I need a bit reader to
branch the gap between them. Ultimately, this was something
that I should have been doing all along.

There are other associated changes in this commit. This
diff is not very clean.