Another Huge commit
This is simply a dump of my work so far. I'm going to put this
to the side for a while. Currently it doesn't work, and probably
doesn't produce the correct output for the audio frames that it
does manage to deframe, so it is still very much a work in
progress.

I'm encountering an issue where I'm failing to read the proper
coefficient precision of the 4th frame (1st subframe). I can't
figure it out, and am thus putting this down for a while to clear
my head.